@@ -5,7 +5,7 @@ class Shot < ActiveRecord::Base
 
   validates_presence_of :name
   validates_presence_of :sequence_id
-  validates_format_of   :name, :with => /^\w{2,3}\d{2,3}\w$/i, 
+  validates_format_of   :name, :with => /\A\w{2,3}\d{2,3}\w\z/i, 
                                :message => 'Shot name must be like: ebc23 or eb023a'
 
   validates_uniqueness_of :name, :scope => [ :sequence_id ], 
